+++ js/prism.js
@@ -117,7 +117,7 @@
         return env.content.split(/\n/).map(function(content) {
             if (content) {
                 return '\x00\x00\x00\x00[[b;;;' + env.classes.join(' ') + ']' +
-                    content + '\x00\x00\x00\x00]';
+                    $.terminal.escape_brackets(content) + '\x00\x00\x00\x00]';
             }
             return '';
         }).join('\n');